backend/requirements.txt
asgiref==3.10.0
certifi==2025.11.12
charset-normalizer==3.4.4
Django==5.2.7
django-ckeditor==6.7.3
django-cors-headers==4.9.0
django-filter==25.2
django-jazzmin==3.0.1
django-js-asset==3.1.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dotenv==0.9.9
idna==3.11
pillow==12.0.0
PyJWT==2.10.1
python-dotenv==1.2.1
requests==2.32.5
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
